refactor(runtime): integrate domain events into runtime infrastructure
- Move agent_resolver from orchestra to runtime module
- Move no_progress_policy from orchestra to runtime module
- Add OrchestraConfig model for event-driven configuration
- Add OrchestraStatusService for status reporting
- Add internal command for event handler registration
- Update CLI and run command to support event-driven mode
- Fix import paths for execute_state_fallback (runtime vs orchestra)
- Add build_lifecycle_callbacks to RunUsecase for event-driven callbacks

Integration points:
- Event handlers registered at runtime initialization
- Agent resolution uses domain events
- State fallback uses domain events
- Status service emits domain events
- RunUsecase generates lifecycle callbacks with domain events
