+++ src/vibe3/orchestra/agent_resolver.py
@@ -1,127 +1,18 @@
-"""Resolve agent options for orchestra-managed roles.
+"""Agent resolver for orchestra roles (backward compatibility layer).
 
-This module provides role-specific agent resolution that is independent
-from the run/plan/review defaults, ensuring manager, supervisor, and
-governance have their own configuration truth.
+This module re-exports functions from runtime layer for backward compatibility.
 """
 
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-from vibe3.agents.backends.codeagent_config import (
-    resolve_effective_agent_options,
-    sync_models_json,
+from vibe3.runtime.agent_resolver import (
+    _resolve_and_sync,
+    resolve_governance_agent_options,
+    resolve_manager_agent_options,
+    resolve_supervisor_agent_options,
 )
-from vibe3.agents.runner import CodeagentExecutionService
-from vibe3.models.review_runner import AgentOptions
 
-if TYPE_CHECKING:
-    from vibe3.config.settings import VibeConfig
-    from vibe3.orchestra.config import OrchestraConfig
-
-
-def _resolve_and_sync(raw_options: AgentOptions) -> AgentOptions:
-    """Resolve preset and sync to ~/.codeagent/models.json."""
-    effective = resolve_effective_agent_options(raw_options)
-    sync_models_json(effective)
-    return effective
-
-
-def resolve_governance_agent_options(config: OrchestraConfig) -> AgentOptions:
-    """Resolve agent options from governance-specific config.
-
-    Args:
-        config: Orchestra configuration with governance settings
-
-    Returns:
-        Effective agent options with preset resolved from config/models.json
-    """
-    gov = config.governance
-    raw_options = AgentOptions(
-        agent=gov.agent,
-        backend=gov.backend,
-        model=gov.model,
-        worktree=False,  # Governance doesn't need worktree
-    )
-
-    return _resolve_and_sync(raw_options)
-
-
-def resolve_supervisor_agent_options(config: OrchestraConfig) -> AgentOptions:
-    """Resolve agent options from supervisor-handoff-specific config.
-
-    Args:
-        config: Orchestra configuration with supervisor handoff settings
-
-    Returns:
-        Effective agent options with preset resolved from config/models.json
-    """
-    handoff = config.supervisor_handoff
-    raw_options = AgentOptions(
-        agent=handoff.agent,
-        backend=handoff.backend,
-        model=handoff.model,
-        worktree=False,  # L2 uses self-managed temporary worktree, no --worktree flag
-    )
-
-    return _resolve_and_sync(raw_options)
-
-
-def resolve_manager_agent_options(
-    config: OrchestraConfig,
-    runtime_config: VibeConfig,
-    worktree: bool = False,
-) -> AgentOptions:
-    """Resolve agent options from orchestra assignee dispatch config.
-
-    This is the shared source of truth for manager role resolution,
-    used by both async runtime and sync CLI execution.
-
-    Resolution order:
-    1. assignee_dispatch.agent          -> preset mode
-    2. assignee_dispatch.backend/model  -> backend direct mode
-    3. run.agent_config                 -> generic fallback
-
-    If assignee_dispatch contains both ``agent`` and ``backend/model``,
-    manager execution intentionally prefers ``agent`` as the primary entry.
-    After that, ``resolve_effective_agent_options`` may resolve the preset into an
-    explicit backend/model, so an additional ``model`` value can still influence
-    the final model selection even when ``agent`` is present.
-
-    Args:
-        config: Orchestra configuration with assignee dispatch settings
-        runtime_config: Vibe runtime config for fallback defaults
-        worktree: Whether to run in worktree mode
-
-    Returns:
-        Effective agent options with preset resolved from config/models.json
-    """
-    ad = config.assignee_dispatch
-    raw_options: AgentOptions
-    if ad.agent or ad.backend:
-        # Use orchestra-specific config when a preset or explicit backend override
-        # is configured for the manager role.
-        raw_options = AgentOptions(
-            agent=ad.agent,
-            backend=ad.backend,
-            model=ad.model,
-            worktree=worktree,
-            timeout_seconds=ad.timeout_seconds,
-        )
-    else:
-        # Fall back to generic run defaults
-        raw_options = CodeagentExecutionService(runtime_config).resolve_agent_options(
-            "run", worktree=worktree
-        )
-        # Override timeout with manager-specific value
-        raw_options = AgentOptions(
-            agent=raw_options.agent,
-            backend=raw_options.backend,
-            model=raw_options.model,
-            worktree=raw_options.worktree,
-            timeout_seconds=ad.timeout_seconds,
-        )
-
-    # Resolve preset mapping from config/models.json and sync to ~/.codeagent
-    return _resolve_and_sync(raw_options)
+__all__ = [
+    "_resolve_and_sync",
+    "resolve_governance_agent_options",
+    "resolve_manager_agent_options",
+    "resolve_supervisor_agent_options",
+]
