+++ src/vibe3/runtime/__init__.py
@@ -1,4 +1,25 @@
-"""Runtime module for Vibe Orchestra.
+"""Orchestra runtime utilities.
 
-Contains event loop, execution engine, and circuit breaker.
+This module provides shared runtime utilities for orchestra, manager, and agents.
+Placed in runtime layer to avoid circular dependencies.
 """
+
+from vibe3.runtime.agent_resolver import (
+    resolve_governance_agent_options,
+    resolve_manager_agent_options,
+    resolve_supervisor_agent_options,
+)
+from vibe3.runtime.no_progress_policy import (
+    has_progress_changed,
+    snapshot_progress,
+)
+
+__all__ = [
+    # no_progress_policy
+    "has_progress_changed",
+    "snapshot_progress",
+    # agent_resolver
+    "resolve_governance_agent_options",
+    "resolve_manager_agent_options",
+    "resolve_supervisor_agent_options",
+]
